Fixed an issue where it switch pass and email when creating account, Converting enum to u32 for server packet id

+++ src/gameloop/sends.rs
@@ -28,7 +28,7 @@ pub fn send_fltalert(
 ) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(128)?;
 
-    buf.write(ServerPackets::FltAlert)?;
+    buf.write(ServerPackets::FltAlert as u32)?;
     buf.write(ftltype)?;
     buf.write(message)?;
     buf.finish()?;
@@ -41,7 +41,7 @@ pub fn send_fltalert(
 pub fn send_loginok(storage: &Storage, socket_id: usize) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(16)?;
 
-    buf.write(ServerPackets::LoginOk)?;
+    buf.write(ServerPackets::LoginOk as u32)?;
     buf.write(storage.time.borrow().hour)?;
     buf.write(storage.time.borrow().min)?;
     buf.finish()?;
@@ -54,7 +54,7 @@ pub fn send_loginok(storage: &Storage, socket_id: usize) -> Result<()> {
 pub fn send_updatemap(world: &World, storage: &Storage, entity: &Entity) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(16)?;
 
-    buf.write(ServerPackets::UpdateMap)?;
+    buf.write(ServerPackets::UpdateMap as u32)?;
     buf.finish()?;
 
     let id: usize = world.get::<&Socket>(entity.0)?.id;
@@ -81,7 +81,7 @@ pub fn send_mapitem(
 
         let mut buf = ByteBuffer::new_packet_with(64)?;
 
-        buf.write(ServerPackets::MapItems)?;
+        buf.write(ServerPackets::MapItems as u32)?;
         buf.write(*item)?;
         buf.write(itemdata)?;
         buf.finish()?;
@@ -99,7 +99,7 @@ pub fn send_mapitem(
 pub fn playerdata(world: &World, _storage: &Storage, entity: &Entity) -> Option<ByteBuffer> {
     let mut buf = ByteBuffer::new_packet_with(512).ok()?;
 
-    buf.write(ServerPackets::PlayerData).ok()?;
+    buf.write(ServerPackets::PlayerData as u32).ok()?;
     buf.write(*entity).ok()?;
     buf.write(
         world
@@ -138,7 +138,7 @@ pub fn send_dir(world: &World, storage: &Storage, entity: &Entity, toself: bool)
     let mut buf = ByteBuffer::new_packet_with(16)?;
     let closure = |toself, id| if toself { Some(id) } else { None };
 
-    buf.write(ServerPackets::PlayerDir)?;
+    buf.write(ServerPackets::PlayerDir as u32)?;
     buf.write(world.get_or_panic::<Dir>(entity).0)?;
     buf.finish()?;
 
@@ -159,7 +159,7 @@ pub fn send_move(world: &World, storage: &Storage, entity: &Entity, warp: bool) 
 
     let pos = world.get_or_panic::<Position>(entity);
 
-    buf.write(ServerPackets::PlayerMove)?;
+    buf.write(ServerPackets::PlayerMove as u32)?;
     buf.write(*entity)?;
     buf.write(pos)?;
     buf.write(world.get_or_panic::<Dir>(entity).0)?;
@@ -183,7 +183,7 @@ pub fn send_data_remove_list(
 ) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(24)?;
 
-    buf.write(ServerPackets::Dataremovelist)?;
+    buf.write(ServerPackets::Dataremovelist as u32)?;
     buf.write(datatype)?;
     buf.write(remove.to_vec())?;
     buf.finish()?;
@@ -202,7 +202,7 @@ pub fn send_data_remove(
 ) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(24)?;
 
-    buf.write(ServerPackets::Dataremove)?;
+    buf.write(ServerPackets::Dataremove as u32)?;
     buf.write(datatype)?;
     buf.write(id)?;
     buf.finish()?;
@@ -215,7 +215,7 @@ pub fn send_data_remove(
 pub fn send_data_remove_all(world: &World, storage: &Storage, id: u64, datatype: u8) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(24)?;
 
-    buf.write(ServerPackets::Dataremove)?;
+    buf.write(ServerPackets::Dataremove as u32)?;
     buf.write(datatype)?;
     buf.write(id)?;
     buf.finish()?;
@@ -231,7 +231,7 @@ pub fn send_vitals(world: &World, storage: &Storage, entity: &Entity) -> Result<
 
     let vitals = world.get_or_panic::<Vitals>(entity);
 
-    buf.write(ServerPackets::PlayerVitals)?;
+    buf.write(ServerPackets::PlayerVitals as u32)?;
     buf.write(vitals.vital)?;
     buf.write(vitals.vitalmax)?;
     buf.finish()?;
@@ -251,7 +251,7 @@ pub fn send_vitals(world: &World, storage: &Storage, entity: &Entity) -> Result<
 pub fn send_inv(world: &World, storage: &Storage, entity: &Entity) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(6500)?;
 
-    buf.write(ServerPackets::PlayerInv)?;
+    buf.write(ServerPackets::PlayerInv as u32)?;
     buf.write(world.cloned_get_or_panic::<Inventory>(entity).items.clone())?;
     buf.finish()?;
 
@@ -264,7 +264,7 @@ pub fn send_inv(world: &World, storage: &Storage, entity: &Entity) -> Result<()>
 pub fn send_invslot(world: &World, storage: &Storage, entity: &Entity, id: usize) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(32)?;
 
-    buf.write(ServerPackets::PlayerInvSlot)?;
+    buf.write(ServerPackets::PlayerInvSlot as u32)?;
     buf.write(id)?;
     buf.write(world.cloned_get_or_panic::<Inventory>(entity).items[id])?;
     buf.finish()?;
@@ -279,7 +279,7 @@ pub fn send_attack(world: &World, storage: &Storage, entity: &Entity, toself: bo
     let mut buf = ByteBuffer::new_packet_with(16)?;
     let closure = |toself, id| if toself { Some(id) } else { None };
 
-    buf.write(ServerPackets::PlayerAttack)?;
+    buf.write(ServerPackets::PlayerAttack as u32)?;
     buf.write(*entity)?;
     buf.finish()?;
 
@@ -298,7 +298,7 @@ pub fn send_attack(world: &World, storage: &Storage, entity: &Entity, toself: bo
 pub fn send_equipment(world: &World, storage: &Storage, entity: &Entity) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(16)?;
 
-    buf.write(ServerPackets::PlayerEquipment)?;
+    buf.write(ServerPackets::PlayerEquipment as u32)?;
     buf.write(world.cloned_get_or_panic::<Equipment>(entity).items.clone())?;
     buf.finish()?;
 
@@ -317,7 +317,7 @@ pub fn send_equipment(world: &World, storage: &Storage, entity: &Entity) -> Resu
 pub fn send_level(world: &World, storage: &Storage, entity: &Entity) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(16)?;
 
-    buf.write(ServerPackets::PlayerLevel)?;
+    buf.write(ServerPackets::PlayerLevel as u32)?;
     buf.write(world.get_or_panic::<Level>(entity).0)?;
     buf.write(world.get_or_panic::<Player>(entity).levelexp)?;
     buf.finish()?;
@@ -330,7 +330,7 @@ pub fn send_level(world: &World, storage: &Storage, entity: &Entity) -> Result<(
 pub fn send_money(world: &World, storage: &Storage, entity: &Entity) -> Result<()> {
     let mut buf = ByteBuffer::new_packet_with(16)?;
 
-    buf.write(ServerPackets::PlayerMoney)?;
+    buf.write(ServerPackets::PlayerMoney as u32)?;
     buf.write(world.get_or_panic::<Money>(entity).vals)?;
     buf.finish()?;
 
@@ -348,7 +348,7 @@ pub fn send_life_status(
     let mut buf = ByteBuffer::new_packet_with(16)?;
     let closure = |toself, id| if toself { Some(id) } else { None };
 
-    buf.write(ServerPackets::PlayerDeath)?;
+    buf.write(ServerPackets::PlayerDeath as u32)?;
     buf.write(*entity)?;
     buf.write(world.get_or_panic::<DeathType>(entity))?;
     buf.finish()?;
@@ -368,7 +368,7 @@ pub fn send_action(world: &World, storage: &Storage, entity: &Entity, toself: bo
     let mut buf = ByteBuffer::new_packet_with(16)?;
     let closure = |toself, id| if toself { Some(id) } else { None };
 
-    buf.write(ServerPackets::PlayerAction)?;
+    buf.write(ServerPackets::PlayerAction as u32)?;
     buf.write(world.get_or_panic::<Dir>(entity).0)?;
     buf.finish()?;
 
@@ -387,7 +387,7 @@ pub fn send_pvp(world: &World, storage: &Storage, entity: &Entity, toself: bool)
     let mut buf = ByteBuffer::new_packet_with(16)?;
     let closure = |toself, id| if toself { Some(id) } else { None };
 
-    buf.write(ServerPackets::PlayerPvp)?;
+    buf.write(ServerPackets::PlayerPvp as u32)?;
     buf.write(world.get_or_panic::<Player>(entity).pvpon)?;
     buf.finish()?;
 
@@ -406,7 +406,7 @@ pub fn send_pk(world: &World, storage: &Storage, entity: &Entity, toself: bool) 
     let mut buf = ByteBuffer::new_packet_with(16)?;
     let closure = |toself, id| if toself { Some(id) } else { None };
 
-    buf.write(ServerPackets::PlayerPk)?;
+    buf.write(ServerPackets::PlayerPk as u32)?;
     buf.write(world.get_or_panic::<Player>(entity).pk)?;
     buf.finish()?;
 
@@ -443,7 +443,7 @@ pub fn send_message(
         MessageChannel::Private => {
             let mut buf = ByteBuffer::new_packet_with(msg.len() + head.len() + 32)?;
 
-            buf.write(ServerPackets::ChatMsg)?;
+            buf.write(ServerPackets::ChatMsg as u32)?;
             buf.write(chan)?;
             buf.write(head)?;
             buf.write(msg)?;
@@ -458,7 +458,7 @@ pub fn send_message(
         MessageChannel::Quest | MessageChannel::Npc => {
             let mut buf = ByteBuffer::new_packet_with(msg.len() + head.len() + 32)?;
 
-            buf.write(ServerPackets::ChatMsg)?;
+            buf.write(ServerPackets::ChatMsg as u32)?;
             buf.write(chan)?;
             buf.write(head)?;
             buf.write(msg)?;
